requirements.txt
Flask>=2.3.3
Flask-SQLAlchemy>=3.0.5
SQLAlchemy>=2.0.25
faiss-cpu>=1.9.0
sentence-transformers>=2.5.0
numpy>=1.24.0
PyPDF2>=3.0.1
python-docx>=1.1.0
python-magic-bin; sys_platform == "win32"
python-magic; sys_platform != "win32"
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
python-telegram-bot